NCI-Agency/anet#2213: Add an Insight showing the positions that have pending assessments
Refactor the PendingAssessmentsNotificationWorker for this; we want to re-use most of the code.
Add an asynchronous Position search so we can use the refactored PendingAssessmentsHelper.
