+++ client/src/components/SearchFilters.js
@@ -495,6 +495,14 @@ export const searchFilters = function() {
           options: ["true", "false"],
           labels: ["Yes", "No"]
         }
+      },
+      "Has Pending Assessments": {
+        component: CheckboxFilter,
+        deserializer: deserializeCheckboxFilter,
+        props: {
+          queryKey: "hasPendingAssessments",
+          msg: "Yes"
+        }
       }
     }
   }
