+++ client/src/components/advancedSearch/CheckboxFilter.js
@@ -4,7 +4,7 @@ import React from "react"
 import { Checkbox, FormGroup } from "react-bootstrap"
 import { deserializeSearchFilter } from "searchUtils"
 
-const CheckboxFilter = ({ asFormField, queryKey, onChange }) => {
+const CheckboxFilter = ({ msg, asFormField, queryKey, onChange }) => {
   const defaultValue = { value: true }
   const toQuery = val => {
     return { [queryKey]: val.value }
@@ -17,7 +17,6 @@ const CheckboxFilter = ({ asFormField, queryKey, onChange }) => {
     toQuery
   )[0]
 
-  const msg = "Authorized for me"
   return !asFormField ? (
     <>{msg}</>
   ) : (
@@ -29,11 +28,13 @@ const CheckboxFilter = ({ asFormField, queryKey, onChange }) => {
   )
 }
 CheckboxFilter.propTypes = {
+  msg: PropTypes.string,
   queryKey: PropTypes.string.isRequired,
   onChange: PropTypes.func, // eslint-disable-line react/no-unused-prop-types
   asFormField: PropTypes.bool
 }
 CheckboxFilter.defaultProps = {
+  msg: "Authorized for me",
   asFormField: true
 }
