+++ src/main/java/mil/dds/anet/AnetApplication.java
@@ -160,7 +160,8 @@ public class AnetApplication extends Application<AnetConfiguration> {
 
     // The Object Engine is the core place where we store all of the Dao's
     // You can always grab the engine from anywhere with AnetObjectEngine.getInstance()
-    final AnetObjectEngine engine = new AnetObjectEngine(dbUrl, this, metricRegistry);
+    final AnetObjectEngine engine =
+        new AnetObjectEngine(dbUrl, this, configuration, metricRegistry);
     environment.servlets().setSessionHandler(new SessionHandler());
 
     if (configuration.isDevelopmentMode()) {
@@ -269,18 +270,19 @@ public class AnetApplication extends Application<AnetConfiguration> {
         new AuthorizationGroupResource(engine);
     final NoteResource noteResource = new NoteResource(engine);
     final ApprovalStepResource approvalStepResource = new ApprovalStepResource(engine);
+    final GraphQlResource graphQlResource = injector.getInstance(GraphQlResource.class);
+    graphQlResource.initialise(engine, configuration,
+        ImmutableList.of(reportResource, personResource, positionResource, locationResource,
+            orgResource, taskResource, adminResource, savedSearchResource,
+            authorizationGroupResource, noteResource, approvalStepResource),
+        metricRegistry);
 
     // Register all of the HTTP Resources
     environment.jersey().register(loggingResource);
     environment.jersey().register(adminResource);
     environment.jersey().register(homeResource);
     environment.jersey().register(new ViewResponseFilter(configuration));
-    environment.jersey()
-        .register(new GraphQlResource(engine, configuration,
-            ImmutableList.of(reportResource, personResource, positionResource, locationResource,
-                orgResource, taskResource, adminResource, savedSearchResource,
-                authorizationGroupResource, noteResource, approvalStepResource),
-            metricRegistry));
+    environment.jersey().register(graphQlResource);
   }
 
   private void runAccountDeactivationWorker(final AnetConfiguration configuration,
