+++ src/main/java/mil/dds/anet/AnetObjectEngine.java
@@ -22,6 +22,7 @@ import mil.dds.anet.beans.Task;
 import mil.dds.anet.beans.search.ISearchQuery.RecurseStrategy;
 import mil.dds.anet.beans.search.OrganizationSearchQuery;
 import mil.dds.anet.beans.search.TaskSearchQuery;
+import mil.dds.anet.config.AnetConfiguration;
 import mil.dds.anet.database.AdminDao;
 import mil.dds.anet.database.AdminDao.AdminSettingKeys;
 import mil.dds.anet.database.ApprovalStepDao;
@@ -73,11 +74,13 @@ public class AnetObjectEngine {
   ISearcher searcher;
 
   private static AnetObjectEngine instance;
+  private static AnetConfiguration configuration;
 
   private final String dbUrl;
   private final Injector injector;
 
-  public AnetObjectEngine(String dbUrl, Application<?> application, MetricRegistry metricRegistry) {
+  public AnetObjectEngine(String dbUrl, Application<?> application, AnetConfiguration config,
+      MetricRegistry metricRegistry) {
     this.dbUrl = dbUrl;
     injector = InjectorLookup.getInjector(application).get();
     personDao = injector.getInstance(PersonDao.class);
@@ -98,6 +101,7 @@ public class AnetObjectEngine {
     jobHistoryDao = injector.getInstance(JobHistoryDao.class);
     this.metricRegistry = metricRegistry;
     searcher = Searcher.getSearcher(DaoUtils.getDbType(dbUrl), injector);
+    configuration = config;
     instance = this;
   }
 
@@ -359,6 +363,10 @@ public class AnetObjectEngine {
 
   public static AnetObjectEngine getInstance() {
     return instance;
+  }
+
+  public static AnetConfiguration getConfiguration() {
+    return configuration;
   }
 
   public String getAdminSetting(AdminSettingKeys key) {
